updated provider_handler
